Add blog-post as submodule
This is an attempt to separate the posting content from the site
generation code. By default Jekyll will look into the _posts folder for
content to convert into web pages.

This is being overwritten by creating a repository for the post content
in "blog-posts" and utilizing the submodule feature of Git to include
these contents in the site generation.

Currently, .gitmodules allows you to redefine the name and path of the
submodule. A simple `mv blog-posts/ _posts/` should trick Jekyll into
converting the contents of _posts into page material file while writing
in the blog-posts repository.