Use JS to open external links in new tab
By default, Markdown parsers do not set links using Markdown syntax
`[]()` to use target='_blank' (opening them in a new tab).

To make links open in a new tab, the target attribute must be set to
_blank. However, links within a blog/domain should not open in new tabs.

A simple handleExternalLinks() function needed to be written to change
external links to open in a new tab, while preserving links within a
domain.